+++ EDDI/ChangeLog.md
@@ -7,11 +7,14 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * Events
       * `Backpack` event added
       * `Backpack changed` event added
+      * `Mission accepted` event variables updated to include micro-resources (on foot items)
+      * `Mission completed` event variables updated to include micro-resources (on foot items)
     * Scripts
       * `Carrier jumped` script, revised to correct missing "Information:" verbiage when the engineer report is the only applicable report.
       * `Engineer report` script, revised to refine grammar.
       * `Jumped` script, revised to correct missing "Information:" verbiage when the engineer report is the only applicable report.
       * `Mission check system` script, fixed a bug that could prevent the station model from being referenced.
+      * `Mission completed` script updated to include micro-resources (on foot items)
 
 ### 4.0.0-b3
   * Core